fix: fix data type error in data classes

+++ app/src/main/java/com/campusmov/uniride/data/datasource/remote/dto/StudentRatingMetricResponseDto.kt
@@ -5,15 +5,15 @@ import com.campusmov.uniride.domain.analytics.model.StudentRatingMetric
 data class StudentRatingMetricResponseDto (
 
     val userId: String?,
-    val totalRatings: Int?,
-    val totalReviewsCount: Double?,
+    val totalRatings: Double?,
+    val totalReviewsCount: Int?,
     val averageRating: Double?
 ){
     fun toDomain(): StudentRatingMetric {
         return StudentRatingMetric(
             userId = userId ?: "",
-            totalRatings = totalRatings ?: 0,
-            totalReviewsCount = totalReviewsCount ?: 0.0,
+            totalRatings = totalRatings ?: 0.0,
+            totalReviewsCount = totalReviewsCount ?: 0,
             averageRating = averageRating ?: 0.0
         )
     }
